+++ pkg/clickhouse/clickhouse.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"path"
 	"path/filepath"
+	"regexp"
 	"strconv"
 	"strings"
 	"syscall"
@@ -634,10 +635,18 @@ func (ch *ClickHouse) CreateTable(table Table, query string, dropTable bool) err
 		return errors.New(fmt.Sprintf("schema query ```%s``` doesn't contains table name `%s`", query, table.Name))
 	}
 
-	// fix restore schema for legacy backup, see https://github.com/AlexAkulov/clickhouse-backup/issues/268
-	if strings.Contains(query, fmt.Sprintf("`%s`", table.Name)) && !strings.Contains(query, fmt.Sprintf("`%s`.`%s`", table.Database, table.Name)) {
+	// fix restore schema for legacy backup, see https://github.com/AlexAkulov/clickhouse-backup/issues/268 and https://github.com/AlexAkulov/clickhouse-backup/issues/297
+	isOnlyTableWithQuotesPresent, err := regexp.Match(fmt.Sprintf("^CREATE [^(]+ `%s`", table.Name), []byte(query))
+	if err != nil {
+		return err
+	}
+	isOnlyTablePresent, err := regexp.Match(fmt.Sprintf("^CREATE [^(]+ %s", table.Name), []byte(query))
+	if err != nil {
+		return err
+	}
+	if isOnlyTableWithQuotesPresent {
 		query = strings.Replace(query, fmt.Sprintf("`%s`", table.Name), fmt.Sprintf("`%s`.`%s`", table.Database, table.Name), 1)
-	} else if strings.Contains(query, table.Name) && !strings.Contains(query, fmt.Sprintf("%s.%s", table.Database, table.Name)) && !strings.Contains(query, table.Database) {
+	} else if isOnlyTablePresent {
 		query = strings.Replace(query, fmt.Sprintf("%s", table.Name), fmt.Sprintf("%s.%s", table.Database, table.Name), 1)
 	}